+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### ✨ Improved
 
 * GORT will fail to initialise if the Overwatcher is running. This can be overridden by passing `override_overwatcher=True` to the `Gort` constructor or `--override-overwatcher` in the CLI.
+* Rearranged the Overwatcher helpers and make the pre- and post-observing scripts recipes.
+* Added a framework for run daily tasks and a pre-observing task.
 
 
 ## 1.0.1 - October 30, 2024